% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/biosys_datasets.R
\name{biosys_datasets}
\alias{biosys_datasets}
\title{Parse BioSys datasets to tibble}
\usage{
biosys_datasets(un = Sys.getenv("BIOSYS_UN"), pw = Sys.getenv("BIOSYS_PW"),
  project_id = NULL)
}
\arguments{
\item{un}{(character) The BioSys username, default \code{Sys.getenv("BIOSYS_UN")}}

\item{pw}{(character) The BioSys password, default \code{Sys.getenv("BIOSYS_PW")}}

\item{project_id}{(int) An optional project ID to restrict the result to.}
}
\value{
A tibble of dataset metadata. The field \code{data_package} is a list
  column containing the \code{tabular_data_package} definition.
}
\description{
Parse BioSys datasets to tibble
}

