% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/biosys_get.R
\name{biosys_get}
\alias{biosys_get}
\title{Return JSON from a BioSys API endpoint}
\usage{
biosys_get(serializer, api = "https://biosys.dbca.wa.gov.au/api/",
  un = Sys.getenv("BIOSYS_UN"), pw = Sys.getenv("BIOSYS_PW"),
  query = list())
}
\arguments{
\item{serializer}{(character) WAStD API serializer name (required).
Possible values as per
\code{https://biosys.dbca.wa.gov.au/api/} are:
\itemize{
\item projects
\item datasets
\item records
\item sites
}}

\item{api}{(character) The BioSys API URL,
default https://biosys.dbca.wa.gov.au/api/}

\item{un}{(character) The BioSys username, default \code{Sys.getenv("BIOSYS_UN")}}

\item{pw}{(character) The BioSys password, default \code{Sys.getenv("BIOSYS_PW")}}

\item{query}{(list) API query parameters for format, limit, filtering
(default: list(taxon='Cheloniidae', limit=10000, format='json'))}
}
\value{
An S3 object of class 'wastd_api_response' containing:

  content: The retrieved JSON as list of lists

  serializer: The called serializer, e.g. 'projects'

  response: The API HTTP response with all metadata
}
\description{
Call the BioSys API serializer's list view with given GET
  parameters and return the response JSON as list of lists.
}

