% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/biosys_projects.R
\name{biosys_projects}
\alias{biosys_projects}
\title{Parse BioSys projects to tibble}
\usage{
biosys_projects(un = Sys.getenv("BIOSYS_UN"), pw = Sys.getenv("BIOSYS_PW"))
}
\arguments{
\item{un}{(character) The BioSys username, default \code{Sys.getenv("BIOSYS_UN")}}

\item{pw}{(character) The BioSys password, default \code{Sys.getenv("BIOSYS_PW")}}
}
\value{
A tibble with project metadata. Fields \code{site_data_package}
  and \code{custodians} are list columns.
}
\description{
Parse BioSys projects to tibble
}

