% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/biosys_records.R
\name{biosys_records}
\alias{biosys_records}
\title{Parse BioSys records to tibble}
\usage{
biosys_records(un = Sys.getenv("BIOSYS_UN"), pw = Sys.getenv("BIOSYS_PW"),
  project_id = NULL)
}
\arguments{
\item{un}{(character) The BioSys username, default \code{Sys.getenv("BIOSYS_UN")}}

\item{pw}{(character) The BioSys password, default \code{Sys.getenv("BIOSYS_PW")}}

\item{project_id}{(int) An optional project ID to restrict the result to.}
}
\value{
A tibble of BioSys records with some common metadata fields plus the
  respective data.
}
\description{
Parse BioSys records to tibble
}

