% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pkg_data.R
\name{records}
\alias{records}
\title{Example records of one dataset}
\format{a \code{tibble} with 100 records of 8 shared and 46 dataset-specific
  variables.}
\description{
Example records of one dataset
}

