% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/biosys_projects.R
\name{biosys_projects}
\alias{biosys_projects}
\title{Parse BioSys projects to tibble}
\usage{
biosys_projects(un = Sys.getenv("BIOSYS_UN"), pw = Sys.getenv("BIOSYS_PW"))
}
\description{
Parse BioSys projects to tibble
}

