% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/biosys_records.R
\name{biosys_records}
\alias{biosys_records}
\title{Parse BioSys records to tibble}
\usage{
biosys_records(un = Sys.getenv("BIOSYS_UN"), pw = Sys.getenv("BIOSYS_PW"),
  project_id = NULL)
}
\description{
Parse BioSys records to tibble
}

